# frontend/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorCode": {
        "description": "Stable HTTP API errors not emitted by a graph run.",
        "enum": [
          "invalid_request",
          "internal_error",
          "handoff_not_pending",
          "tool_approval_not_pending",
          "session_already_exists",
          "session_busy",
          "session_not_found",
          "knowledge_unavailable"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
...
      "Body_upload_file_files_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_file_files_post",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "description": "A sanitized stable API error.",
        "properties": {
          "error_code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "FastAPI's standard error envelope.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FileUploadResponse": {
        "additionalProperties": false,
        "description": "文件上传回执(D7-T1):url 为受控下载的相对路径。\n\n- file_id 是服务端生成的 uuid4().hex + 白名单后缀(落盘名),url 形如\n  /files/{file_id}——客户端凭 url 即可 GET 下载,url 不含原始文件名\n  (后者只作展示字段 name);\n- content_type 由服务端按扩展名映射,不信任客户端伪造的类型;\n- name 是原始文件名(仅展示用;落盘名是 uuid,见 api/files.py 的\n  防穿越设计)。",
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "file_id": {
            "title": "File Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "file_id",
          "name",
          "content_type",
          "size",
          "url"
        ],
        "title": "FileUploadResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/files": {
      "post": {
        "description": "上传聊天附件(按用户隔离存储),返回受控下载回执。\n\n- 扩展名白名单 / 大小上限 / 空文件都在 API 层拦截为 422(大小在\n  逐块读取时累计,不能信 Content-Length);\n- 存储目录按 user_key 隔离(消毒后的 X-User-Id,None → anonymous);\n  落盘名是 uuid4().hex + 白名单后缀,原始文件名只作展示字段返回;\n- 同步写盘走 run_in_threadpool,不阻塞事件循环;写盘失败(OSError\n  等)统一映射 500 internal_error,不泄底层细节。",
        "operationId": "upload_file_files_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_files_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Upload File",
        "tags": [
          "files"
        ]
      }
    },
    "/files/{file_id}": {
      "get": {
        "description": "按受控下载路径提供已上传文件(用户隔离校验)。\n\n- file_id 必须是安全段(见 _is_safe_segment),不匹配一律 404;\n- 存储路径 {root}/{user_key}/{file_id}:user_key 由当前 X-User-Id\n  消毒得出(与上传同源)——他人目录下的 uuid 文件名不可枚举,\n  等价于文件不存在,统一 404(不返回 403,避免泄露目录存在性);\n- media_type 按扩展名映射,与上传响应的 content_type 同源\n  (_UPLOAD_CONTENT_TYPES);底层读盘异常映射 500。",
        "operationId": "get_uploaded_file_files__file_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "file_id",
            "required": true,
            "schema": {
              "title": "File Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Uploaded File",
        "tags": [
          "files"
        ]
      }
    },
...
  },
...
}